+++ crates/sandogasa-inventory/src/content_resolver.rs
@@ -162,6 +162,7 @@ mod tests {
                     repology_name: None,
                     distros: None,
                     file_issue: None,
+                    priority: None,
                 },
                 Package {
                     name: "systemd".to_string(),
@@ -184,6 +185,7 @@ mod tests {
                     repology_name: None,
                     distros: None,
                     file_issue: None,
+                    priority: None,
                 },
                 Package {
                     name: "neovim".to_string(),
@@ -197,6 +199,7 @@ mod tests {
                     repology_name: None,
                     distros: None,
                     file_issue: None,
+                    priority: None,
                 },
             ],
         }